Fix the proxy: no tunnel was ever built, and none kept its real host
Every proxied connection failed at the first step. The loopback port for a
tunnel is derived with FNV-1a, which is defined over arithmetic modulo 2^32 and
relies on the multiply wrapping around - but Ui.csproj sets
CheckForOverflowUnderflow, so the whole assembly compiles checked and the very
first round threw OverflowException. ApplyTo caught it, offered to fall back to a
direct connection, and no tunnel was ever created. The hash now opts out of
checked arithmetic explicitly.

Pointing a session at a tunnel also threw its real endpoint away, because ApplyTo
overwrote Address and Port in place. Two things broke outright. BuildConnectionId
is computed from the address, so the lookup that runs before the proxy is applied
could never find the host registered after it, and a second click opened a
duplicate RDP or VNC session instead of activating the one already open. And an
RDP reconnect that takes the alternate-address path called GetCredential, which
refreshes the server from the data source and so put the real address back,
sending the reconnect around the proxy. A session now remembers where it was
really going: Address and Port still name the tunnel, because that is what the
runner has to dial, while RealAddress and RealPort feed everything that
identifies or describes it - the connection id, the subtitle behind the
fullscreen title and the RDP toolbar, the tab's copy-address and RDP export, and
the SERVER_HOST handed to the after-disconnect script, which until now disagreed
with the one the before-connect script saw.

Proxy passwords were kept enciphered in the very property the settings editor is
bound to, so the field displayed a block of cipher text - and an empty password
was enciphered too, which is why a proxy with no password showed noise. They are
plain in memory now and enciphered only on the way into the profile, the way the
database sources already do it, and the editor uses the same masked input as
every other password field. A pooled tunnel also adopts the current password
whenever it is handed out, so correcting a mistyped one no longer needs a
restart. Proxy addresses are trimmed.

CheckForOverflowUnderflow turned out to bite in four more places, all reachable.
A failure HRESULT converted to int raised OverflowException instead of the
COMException the caller is written to handle, in all six of WindowsShortcut's COM
calls. A password containing CJK or an emoji broke reading credentials back out
of the Windows prompt, because Marshal.ReadInt16 is signed. A transfer whose
progress ran past the size the remote declared underflowed an unsigned
subtraction while rendering the remaining time. And an icon taller than it is
wide scaled past what an ICO directory entry can describe. The acrylic tint pack
is hardened the same way, before its alpha ever reaches 0x80.

FTP was the only address-based protocol with no proxy picker, because its editor
hardcoded the host and port instead of using the shared host view, even though
ApplyTo has always been able to tunnel it.

+++ Ui/Utils/Theme/AcrylicHelper.cs
@@ -139,8 +139,10 @@ namespace _1RM.Utils.Theme
             {
                 AccentState = state,
                 AccentFlags = state == AccentState.Disabled ? 0 : DRAW_ALL_BORDERS,
-                // the struct wants 0xAABBGGRR, which is byte-reversed from WPF's ARGB
-                GradientColor = (uint)((tint.A << 24) | (tint.B << 16) | (tint.G << 8) | tint.R),
+                // the struct wants 0xAABBGGRR, which is byte-reversed from WPF's ARGB. Each byte is widened to
+                // uint before it is shifted: packing in int instead makes any alpha of 0x80 or more negative,
+                // and this assembly's checked arithmetic rejects that on the way back to uint.
+                GradientColor = ((uint)tint.A << 24) | ((uint)tint.B << 16) | ((uint)tint.G << 8) | tint.R,
                 AnimationId = 0,
             };
